APP/requirements.txt
asgiref==3.7.2
dj-static==0.0.6
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
greenlet==3.0.3
numpy==1.26.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.9.0
tzdata==2023.3
